refactor(cosh-ng): narrow shell insight module boundaries
