+++ src/cosh-ng/crates/cosh-shell/src/types/public.rs
@@ -1,16 +1,18 @@
+#[allow(dead_code)]
 #[path = "mod.rs"]
 mod implementation;
 
 pub use implementation::{
-    AgentEvent, AgentMode, AgentRequest, AuditRecord, BuiltinFindingFacts, CommandBlock,
-    CommandOrigin, CommandStatus, CoshApprovalMode, EvaluatedHookFinding, Finding, FindingKind,
-    FindingSeverity, GovernanceDecision, GovernancePolicyDecision, GovernedEvent,
-    HighMemoryProcessFacts, HookFinding, HookProvenance, Intervention, InterventionDecision,
-    MemoryPressureFacts, MetricsConfidence, OutputRefs, Policy, ProcessMemoryFact,
-    QuestionSelectionMode, ShellEnvironmentSnapshot, ShellEvent, ShellEventKind,
-    ShellHandoffRequest, COMMAND_OUTPUT_REF_MAX_BYTES, SESSION_OUTPUT_REF_MAX_BYTES,
+    AgentEvent, AgentMode, AgentRequest, AuditRecord, CommandBlock, CommandOrigin, CommandStatus,
+    CoshApprovalMode, Finding, FindingKind, FindingSeverity, GovernanceDecision,
+    GovernancePolicyDecision, GovernedEvent, HookFinding, Intervention, InterventionDecision,
+    OutputRefs, Policy, QuestionSelectionMode, ShellEvent, ShellEventKind, ShellHandoffRequest,
+    COMMAND_OUTPUT_REF_MAX_BYTES, SESSION_OUTPUT_REF_MAX_BYTES,
 };
 
+#[allow(unused_imports)]
 pub(crate) use implementation::{
-    set_request_context_binding, AgentContextBinding, BuiltinFactRecord,
+    set_request_context_binding, AgentContextBinding, BuiltinFactRecord, BuiltinFindingFacts,
+    EvaluatedHookFinding, HighMemoryProcessFacts, HookProvenance, MemoryPressureFacts,
+    MetricsConfidence, ProcessMemoryFact, ShellEnvironmentSnapshot,
 };
